Finish CheckExistant

+++ src/main/java/ch/neeman/guy/secretmessage/Controller.java
@@ -1,6 +1,7 @@
 package ch.neeman.guy.secretmessage;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cglib.core.Local;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.web.bind.annotation.*;
@@ -31,8 +32,13 @@ public class Controller {
         String message = request.get("message");
         String password = encoder.encode(request.get("password"));
 
-        LocalDateTime expireAt = LocalDateTime.now().plusMinutes(Integer.parseInt(request.get("expirationminute")));
         Boolean selfdelete = Boolean.valueOf(request.get("selfdelete"));
+        LocalDateTime expireAt;
+        if (selfdelete) {
+            expireAt = null;
+        } else {
+            expireAt = LocalDateTime.now().plusMinutes(Integer.parseInt(request.get("expirationminute")));
+        }
         String uuid = UUID.randomUUID().toString();
 
         SecretMessage msg = new SecretMessage(
@@ -72,10 +78,15 @@ public class Controller {
     public ResponseEntity<String> checkExistance(@PathVariable String uuid) {
         return repo.findByUuid(uuid)
                 .map(secretMessage -> {
+                    if (secretMessage.getExpireAt() == null) {
+                        return ResponseEntity.ok("Message exists (self-delete enabled)!");
+                    }
+
                     if (LocalDateTime.now().isAfter(secretMessage.getExpireAt())) {
                         repo.delete(secretMessage);
                         return ResponseEntity.status(410).body("Message has expired!");
                     }
+
                     return ResponseEntity.ok("Message exists!");
                 })
                 .orElseGet(() -> ResponseEntity.status(404).body("Message not found"));
